db rewrite
